pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.111.0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.32",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.6",
    "httpx>=0.27.2",
    "aiosqlite>=0.19.0",
]

